RCDY-286 feat: fetch movie credits
영화 배우 정보, 감독 정보를 가져온다.

+++ service-movie/impl/src/test/java/com/recody/recodybackend/movie/features/getmoviecredit/TMDBGetMovieCreditHandlerTest.java
@@ -9,14 +9,14 @@ import org.springframework.test.context.ContextConfiguration;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @ActiveProfiles("test")
 @ContextConfiguration(classes = RecodyMovieApplication.class)
 class TMDBGetMovieCreditHandlerTest {
     
-    public static final long MOVIE_ID = 705996L;
+    public static final Integer MOVIE_ID = 705996;
     @Autowired
     private GetMovieCreditHandler getMovieCreditHandler;
